+++ src/controllers/leadControllers.js
@@ -1,34 +1,20 @@
 const LeadModel = require('../models/leadSchema')
 const UserModel = require('../models/userSchema')
-const leadsUtils = require('../utils/leads')
+const { paginatedQuery } = require('../utils/queries')
+const { queryBuilderToGetLeads } = require('../utils/leads')
 
 const controller = {}
 
 controller.getLeads = async (req, resp) => {
     let { limit, page } = req.query
+
     limit = Number.parseInt(limit)
     page = Number.parseInt(page)
 
-    const results = {}
-    results.next = null
     try {
-        const total = await LeadModel.find({})
-        const leads = await LeadModel.find({})
-            .limit(limit || 5)
-            .skip((page - 1) * limit || 0)
-
-        const next = await LeadModel.find({})
-            .limit(limit || 5)
-            .skip(page * limit || 5)
-
-        if (next.length > 0) {
-            results.next = `${process.env.HOST}/leads?page=${page + 1 || 2}&limit=${limit || 5}`
-        }
-
-        results.total = total.length //Temp resolution
-        results.results = await leadsUtils.queryBuilderToGetLeads(leads)
-        resp.status(200).json(results)
+        paginatedQuery(LeadModel, limit, page, queryBuilderToGetLeads, resp)
     } catch (err) {
+        console.log(err)
         resp.status(500).json({ error: "Ocorreu um erro ao recuperar as leads" })
     }
 }
